backend/requirements.txt
psycopg2-binary
nltk
vaderSentiment
praw
